Support plugins for DB (base, elastic, mongo)
